xea-core/requirements.txt
cffi==1.7.0
cryptography==1.4
Django==1.9.8
django-cors-headers>=1.1.0,<2.0.0
django-debug-toolbar==1.5
django-filter==0.13.0
django-rest-knox==2.2.0
djangorestframework==3.4.0
idna==2.1
PyJWT>=1.4.0,<2.0.0
Markdown==2.6.6
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==16.0.0
six==1.10.0
sqlparse==0.2.0